refactor: collaborators page, added metadata
